Flytt mapping til PeriodeResultat
