+++ src/test/kotlin/no/nav/familie/ba/sak/behandling/PeriodeMapperTest.kt
@@ -4,8 +4,7 @@ import no.nav.familie.ba.sak.behandling.fagsak.FagsakService
 import no.nav.familie.ba.sak.behandling.vilkår.PersonResultat
 import no.nav.familie.ba.sak.behandling.vilkår.Vilkår
 import no.nav.familie.ba.sak.behandling.vilkår.VilkårResultat
-import no.nav.familie.ba.sak.beregning.domene.PeriodeResultat
-import no.nav.familie.ba.sak.beregning.domene.PeriodeVilkår
+import no.nav.familie.ba.sak.beregning.domene.foldTidslinjer
 import no.nav.familie.ba.sak.common.DbContainerInitializer
 import no.nav.familie.ba.sak.common.lagBehandling
 import no.nav.familie.ba.sak.common.lagBehandlingResultat
@@ -14,7 +13,6 @@ import no.nav.familie.ba.sak.config.ApplicationConfig
 import no.nav.fpsak.tidsserie.LocalDateInterval
 import no.nav.fpsak.tidsserie.LocalDateSegment
 import no.nav.fpsak.tidsserie.LocalDateTimeline
-import no.nav.fpsak.tidsserie.StandardCombinators
 import no.nav.nare.core.evaluations.Resultat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
@@ -36,50 +34,13 @@ import java.time.LocalDate
 @ActiveProfiles("postgres", "mock-dokgen", "mock-oauth")
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @AutoConfigureWireMock(port = 28085)
-class MapperT {
+class PeriodeMapperTest {
 
     @Autowired
     lateinit var behandlingService: BehandlingService
 
     @Autowired
     lateinit var fagsakService: FagsakService
-
-    private fun kombinerVerdier(lhs: LocalDateTimeline<List<VilkårResultat>>,
-                                rhs: LocalDateTimeline<VilkårResultat>): LocalDateTimeline<List<VilkårResultat>> {
-        return lhs.combine(rhs,
-                           { dateInterval, left, right ->
-                               StandardCombinators.allValues<VilkårResultat>(dateInterval,
-                                                                             left,
-                                                                             right)
-                           },
-                           LocalDateTimeline.JoinStyle.CROSS_JOIN)
-    }
-
-    private fun foldTidslinjer(tidslinjer: List<LocalDateTimeline<VilkårResultat>>): LocalDateTimeline<List<VilkårResultat>> {
-        val førsteSegment = tidslinjer.first().toSegments().first()
-        val initiellSammenlagt =
-                LocalDateTimeline(listOf(LocalDateSegment(førsteSegment.fom, førsteSegment.tom, listOf(førsteSegment.value))))
-        val resterende = tidslinjer.drop(1)
-        return resterende.fold(initiellSammenlagt) { sammenlagt, neste -> (kombinerVerdier(sammenlagt, neste)) }
-    }
-
-    fun personResultatTilPeriodeResultater(personResultat: PersonResultat): List<PeriodeResultat> {
-        val tidslinjer = personResultat.vilkårResultater.map { vilkårResultat ->
-            LocalDateTimeline(listOf(LocalDateSegment(vilkårResultat.periodeFom,
-                                                      vilkårResultat.periodeTom,
-                                                      vilkårResultat)))
-        }
-        val kombinertTidslinje = foldTidslinjer(tidslinjer)
-        val periodeResultater = kombinertTidslinje.toSegments().map { segment ->
-            PeriodeResultat(
-                    personIdent = personResultat.personIdent,
-                    periodeFom = segment.fom,
-                    periodeTom = segment.tom,
-                    vilkårResultater = segment.value.map { PeriodeVilkår(it.vilkårType, it.resultat, it.begrunnelse) }.toSet()
-            )
-        }
-        return periodeResultater
-    }
 
     @Test
     fun `Kombinert tidslinje returnerer rette rette vilkårsresultater for tidsintervaller`() {
@@ -132,8 +93,6 @@ class MapperT {
         val d3 = LocalDate.now()
         val d4 = LocalDate.now().plusMonths(1)
         val d5 = LocalDate.now().plusMonths(2)
-
-
         val fnr1 = randomFnr()
         val fnr2 = randomFnr()
         val fagsak = fagsakService.hentEllerOpprettFagsakForPersonIdent(fnr1)
@@ -159,8 +118,8 @@ class MapperT {
                                                                 periodeFom = d2,
                                                                 periodeTom = d5,
                                                                 begrunnelse = ""))
-        val personResultater = listOf(personResultat1, personResultat2)
-        val periodeResultater = personResultater.map { personResultatTilPeriodeResultater(it) }.flatten()
+        behandlingResultat.personResultater = setOf(personResultat1, personResultat2)
+        val periodeResultater = behandlingResultat.periodeResultater.toList()
 
         assert(periodeResultater.size == 4)
